Move sys_model_task to _GeneralModuleFiles
Upcoming mujoco features make use of SysModelTask
